add extractor logging back, and fix collection logging. Fixes #183
After adding normal logging to the docserver logger we left the
database handler commented out. We restore it, and give it another
name so that it doesn't conflict.
Also fix a copy-paste error when getting the hostname in
process_collection - giving all items in a collection to an
extractor in a single go